파일분리 (20.12.01)
파일분리 재시도
보드 > 리스트 , 리스트제외
