+++ TeamProject/static/js/board/main.js
@@ -331,3 +331,18 @@ export const file_dataHub = class {
 }
 
 export const INPUT_DATA_FILE = new file_dataHub();
+//===========나연 남길거============//
+
+/*=============================사이드바 =========================*/
+// 베스트 게시글 불러오기
+export async function loading_best_post() {
+  try {
+    const board_id = location.hash.split('#')[1];
+    const data = await FETCH.get_best_post_information(board_id);
+    if (data != null) {
+      REND.best_post(data);
+    }
+  } catch (error) {
+    console.log(error);
+  }
+}
