+++ TeamProject/static/js/board/router.js
@@ -4,6 +4,7 @@ import * as ASIDE from "./aside/main.js";
 import * as LIST from "./list/index.js"
 import * as EVENT_LIST from "./list/event.js"
 import * as REND_LIST from "./list/render.js"
+import * as RENDER from "./render.js"
 
 /*===========URL 라우팅 형식=========
 게시판 메인화면 : /post#board_id#postmain
@@ -26,25 +27,25 @@ async function router() {
     const hashValue = location.hash.split('#');
     const router_map = {
       postmain: function () { //게시판별 메인페이지
-        REND_LIST.title_and_side_search_setting(hashValue);
+        RENDER.title_and_side_search_setting(hashValue);
         LIST.loading_post(hashValue);
-        ASIDE.loading_best_post();
+        MAIN.loading_best_post();
         EVENT.handle_search();
         window.addEventListener('scroll', EVENT_LIST.handle_scrollHeight);
         return 'postmain';
       },
       postinfo: function () { //게시글 크게보기
         window.removeEventListener('scroll', EVENT_LIST.handle_scrollHeight);
-        REND_LIST.title_and_side_search_setting(hashValue);
+        RENDER.title_and_side_search_setting(hashValue);
         MAIN.load_postinfo(hashValue);
-        ASIDE.loading_best_post();
+        MAIN.loading_best_post();
         EVENT.handle_search();
         return 'postinfo';
       },
       search: function () {
         LIST.loading_search_result(hashValue); //전체게시판검색이면 board_id가 total\
-        ASIDE.loading_best_post();
-        EVENT_LIST.attach_event_when_title_click();
+        MAIN.loading_best_post();
+        EVENT.attach_event_when_title_click();
         EVENT.handle_search();
         window.addEventListener('scroll', EVENT_LIST.handle_scrollHeight);
         return 'search';
